+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- Always add `stackTrace` for `captureException`, using `StackTrace.current` if
+  unset.
+
 ## 0.3.2
 
 - Fix detection of `flutter_driver` for events other than the first.